pyproject.toml
[project]
name = "byoa-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.17.0",
    "litellm>=1.79.0"
]
